requirements.txt
yfinance
pandas
openpyxl
azure-functions
python-dotenv
pyodbc
python-jose[cryptography]
pymupdf
sqlmodel
bcrypt==4.0.1
sqlalchemy==2.0.44
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.0
ijson
fastapi==0.115.6
uvicorn[standard]==0.32.1
gunicorn==23.0.0
python-multipart==0.0.9